Add member-wise initializers to vatom, face, and action Convert blockv region to use member-wise initializers
