+++ BlockV/Core/Network/Models/Package/Vatom/VatomResourceModel.swift
